requirements.txt
aws-cdk-lib==2.117.0
constructs==10.0.0